move extension hook tests into test_vllm_integration.py
